Linear-Regression-Problem-Project
Using Linear Regression to predict the amount of gallons required for a car that travels
1200 miles. A dataset is provided that contains the record of the distance the car has travelled in the past
and the amount of fuel required for each journey. Model a Linear Regression classifier and show the Regression Line
and the goodness of fit, as well as predict the value for the dependent variable using both, the model and Sklearn
module.